Getting ready for realease of 14.10.In this version I made some changes.
Exchanged SimpleScreenRecorder with Kazam Screencaster
Exchanged Google Chrome with Chromium
Temporarily removed icedtea-7-plugin as it has a bug in 14.10
Added an autostart for the hddtemp daemon on boot
Dropbox gets added, but will not install properly on 14.10 yet

This one is only tested on Ubuntu 14.10. The rest will be tested over the last couple of weeks.